+++ host/lib/host_linux.go
@@ -49,7 +49,7 @@ func LocalRun(ctx context.Context, cmd types.Cmd) (types.WaitStatus, error) {
 	}
 
 	waitStatus := types.WaitStatus{}
-	waitStatus.ExitCode = execCmd.ProcessState.ExitCode()
+	waitStatus.ExitCode = uint32(execCmd.ProcessState.ExitCode())
 	waitStatus.Exited = execCmd.ProcessState.Exited()
 	signal := execCmd.ProcessState.Sys().(syscall.WaitStatus).Signal()
 	if signal > 0 {
